some progress debugging chatbot, still a bit more to go

+++ project2/model.py
@@ -73,7 +73,7 @@ class Seq2SeqModel(object):
         self.vocab_size = vocab_size
         self.batch_size = batch_size
         self.learning_rate = tf.Variable(
-            float(learning_rate), trainable=False, dtype=dtype)
+            learning_rate, trainable=False, dtype=tf.float32)
         self.learning_rate_decay_op = self.learning_rate.assign(
             self.learning_rate * learning_rate_decay_factor)
         self.global_step = tf.Variable(0, trainable=False)
